requirements.lock.txt
accelerate==0.22.0
certifi==2022.12.7
charset-normalizer==2.1.1
coloredlogs==15.0.1
diffusers==0.22.3
fast-histogram==0.14
filelock==3.20.0
flatbuffers==25.12.19
fsspec==2025.12.0
huggingface-hub==0.25.2
humanfriendly==10.0
idna==3.4
importlib_metadata==8.7.1
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.6.1
numpy==1.26.4
onnx==1.16.1
onnxoptimizer==0.3.8
onnxruntime==1.23.2
opencv-python==4.11.0.86
packaging==25.0
pillow==12.0.0
protobuf==3.20.3
psutil==7.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.28.1
rknn-toolkit2==2.3.2
ruamel.yaml==0.19.0
ruamel.yaml.clibz==0.3.4
safetensors==0.7.0
scipy==1.16.3
sympy==1.14.0
tokenizers==0.19.1
torch==2.1.2
torchvision==0.16.2
tqdm==4.67.1
transformers==4.41.2
typing_extensions==4.15.0
urllib3==1.26.13
zipp==3.23.0
Flask==3.0.3
Werkzeug==3.0.3
itsdangerous==2.2.0
click==8.1.7
blinker==1.8.2
gfpgan==1.3.8
opencv-python==4.11.0.86
rknn-toolkit-lite2==2.3.2
